requirements.txt
temporalio>=1.6.0
vaderSentiment>=3.3.2
python-dotenv>=1.0.0
